[actions] Extract to a mini class
